remove ImageIO dependencies from Poisson
compiles with clang++ but not with g++
